[sailfish-components-gallery] Fix images sometimes not showing. Contributes to JB#14271
Use ImageMetadata as a fallback for maximumWidth, maximumHeight and
orientation if no value or no valid value is bound to those properties.
Since reading the data from files and creating file system watchers are
both expensive doing either in the general case can contribute to
animation glitches when panning between items so try and avoid doing
either if the data is unneeded.