[components-gallery] Avoid overlapping error messages. Fixes JB#58127
When the user tries to edit an image, but that image no longer exists
(e.g. they press on the "Edit" button of a screenshot notification after
the image was already deleted elsewhere), there are two overlapping
error messgages displayed, as well as an overlapping spinner.

This change ensures only one of the error messages is displayed, and
that the spinner no longer shows while the error is displaying.